+++ packages/connect-kit/src/components/ExtensionInstallModal/ExtensionInstallModal.tsx
@@ -4,18 +4,18 @@ import {
   ModalSection,
   ModalTitle,
 } from "../Modal/Modal.styles";
-import { ConnectFeature } from "./ExtensionUnavailableModal.styles";
+import { ConnectFeature } from "./ExtensionInstallModal.styles";
 import { default as LightbulbSvg } from "../../assets/svg/Lightbulb.svg";
 import { default as CheckmarkSvg } from "../../assets/svg/Checkmark.svg";
 import { getDebugLogger } from "../../lib/logger";
 
-const log = getDebugLogger('ExtensionUnavailableModal')
+const log = getDebugLogger('ExtensionInstallModal')
 
-export type ExtensionUnavailableModalProps = ModalProps;
+export type ExtensionInstallModalProps = ModalProps;
 
-const ExtensionUnavailableModal = ({
+const ExtensionInstallModal = ({
   onClose = () => void 0,
-}: ExtensionUnavailableModalProps) => {
+}: ExtensionInstallModalProps) => {
   log('initializing');
 
   const APP_STORE_URL = 'https://apps.apple.com/app/ledger-extension-browse-web3/id1627727841';
@@ -45,4 +45,4 @@ const ExtensionUnavailableModal = ({
   );
 };
 
-export default ExtensionUnavailableModal;
+export default ExtensionInstallModal;