+++ packages/connect-kit/src/providers/ExtensionSolana.ts
@@ -2,13 +2,13 @@ import { ProviderNotFoundError } from "../lib/errors";
 import { ProviderResult } from "../lib/provider";
 import { getDebugLogger } from "../lib/logger";
 
-const log = getDebugLogger('LedgerConnectSolana');
+const log = getDebugLogger('ExtensionSolana');
 
-export const LEDGER_SOLANA_PROVIDER = 'solana'
-export const LEDGER_CONNECT_SOLANA_PROP = 'isLedgerConnect'
+export const EXTENSION_SOLANA_PROVIDER = 'solana'
+export const EXTENSION_SOLANA_PROP = 'isLedgerConnect'
 
 export interface SolanaProvider {
-  [LEDGER_CONNECT_SOLANA_PROP]: boolean;
+  [EXTENSION_SOLANA_PROP]: boolean;
   signTransaction(...args: unknown[]): Promise<unknown>;
   signAllTransactions(...args: unknown[]): Promise<unknown>;
   signAndSendTransaction(...args: unknown[]): Promise<unknown>;
@@ -17,17 +17,20 @@ export interface SolanaProvider {
 }
 
 interface WindowWithSolana {
-  [LEDGER_SOLANA_PROVIDER]?: SolanaProvider;
+  [EXTENSION_SOLANA_PROVIDER]?: SolanaProvider;
 }
 
+/**
+ * Gets the Solana provider.
+ */
 export function getSolanaProvider (): ProviderResult {
   log('getSolanaProvider');
 
-  const provider = (window as WindowWithSolana)[LEDGER_SOLANA_PROVIDER];
+  const provider = (window as WindowWithSolana)[EXTENSION_SOLANA_PROVIDER];
 
   if (
     typeof provider === "undefined" ||
-    typeof provider[LEDGER_CONNECT_SOLANA_PROP] === "undefined"
+    typeof provider[EXTENSION_SOLANA_PROP] === "undefined"
   ) {
     throw new ProviderNotFoundError();
   }